+++ R/veg_starts.R
@@ -6,8 +6,8 @@
 #
 # Start of vegetation periode according to Menzel (1997)
 #
-# - counting cold days in Nov & dec of previous year
-# - count cold days in current year
+# - counting chill days in Nov & dec of previous year
+# - count chill days in current year
 # - determine critical temperature per days using regression equation fitted
 #   by Menzel 1997
 # - calculate HeatSum starting from Feb
@@ -19,19 +19,24 @@
 #   Phänologischen Gärten und Möglichkeiten der Modellierung von Phänodaten.
 #   - Forstliche Forschungsberichte München, 164, 147 S.
 #==============================================================================
-.start_menzel <- function(df, first.avg=0,
-                         species=c("Larix decidua", "Picea abies (frueh)",
-                                   "Picea abies (spaet)",
-                                   "Picea abies (noerdl.)", "Picea omorika",
-                                   "Pinus sylvestris", "Betula pubescens",
-                                   "Quercus robur", "Quercus petraea",
-                                   "Fagus sylvatica")){
+.start_menzel <- function(df, est.prev, species){
   # Assumptions:
   # - data.frame 'df' contains month, DOY, Tavg
-  # - time range at least till DOY 279
+  # - full years
+  # - last year at least till DOY 279
+  # - and previous Nov & Dec if est.prev==0
 
+  possible.species <- c("Larix decidua", "Picea abies (frueh)",
+                        "Picea abies (spaet)", "Picea abies (noerdl.)",
+                        "Picea omorika", "Pinus sylvestris",
+                        "Betula pubescens", "Quercus robur", "Quercus petraea",
+                        "Fagus sylvatica")
+  if(missing(species))
+    stop(paste0("Argument 'species' is missing. It must be one of '",
+               paste(possible.species, collapse="', '"), "'."))
+
+  species <- match.arg(species, choices=possible.species)
   years <- unique(df$year)
-  species <- match.arg(species)
 
   # Menzel's Parameter
   Mp <- switch(species,
@@ -46,36 +51,36 @@
                "Quercus petraea"       = list(TbCD=9, TbH=3, a=1741, b=-282),
                "Fagus sylvatica"       = list(TbCD=9, TbH=6, a=1922, b=-348))
 
-  # Cold Days
+  # Chill Days
   #----------------------------------------------------------------------------
-  # number of cold days in Nov and Dec trigger next years bud burst
+  # number of chill days in Nov and Dec trigger next years bud burst
   CDNovDec <- stack(tapply(df[(df$month >= 11 & df$Tavg <= Mp$TbCD), "Tavg"],
                            df[(df$month >= 11 & df$Tavg <= Mp$TbCD), "year"],
                            FUN=length))
 
-  # shove the cold days in to next year, where they trigger veg start
+  # shove the chill days in to next year, where they trigger veg start
   CDNovDec$ind <- as.integer(levels(CDNovDec$ind)) + 1
   names(CDNovDec) <- c("CDprev", "year")
 
   # handle first year
-  if(first.avg == 0){
+  if(est.prev == 0){
     # drop first year
-    df[df$years != years[1]-1, ]
+    df <- df[df$year != years[1], ]
   } else {
-    # mean of "first.avg number of years" as proxy for first year's previous cold days
-    CDNovDec <- rbind(c(mean(CDNovDec$CDprev[first.avg]), years[1]),
+    # mean of "est.prev number of years" as proxy for first year's previous chill days
+    CDNovDec <- rbind(c(mean(CDNovDec$CDprev[est.prev]), years[1]),
                       CDNovDec)
   }
 
-  # cumulative sums of cold days per year
+  # cumulative sums of chill days per year
   df$CD <- ifelse(df$Tavg <=  Mp$TbCD, 1, 0)
-  ColdDays <- stack(tapply(df$CD, df$year, FUN=cumsum))
-  names(ColdDays) <- c("CDcumsum", "year")
+  ChillDays <- stack(tapply(df$CD, df$year, FUN=cumsum))
+  names(ChillDays) <- c("CDcumsum", "year")
 
   # merge and add to workhorse df
-  ColdDays <- merge(ColdDays, CDNovDec, by="year", all.x=TRUE)
-  df$CD <- ColdDays$CDprev +  ColdDays$CDcumsum
-  rm(ColdDays, CDNovDec)
+  ChillDays <- merge(ChillDays, CDNovDec, by="year", all.x=TRUE)
+  df$CD <- ChillDays$CDprev +  ChillDays$CDcumsum
+  rm(ChillDays, CDNovDec)
 
   # determine vegetation start with Menzel's regression
   #----------------------------------------------------------------------------
@@ -96,33 +101,42 @@
   return(start)
 }
 
+
 #==============================================================================
 #
-# Start of vegetation periode according to standard climatology procedure
+# Start of vegetation periode according to standard meteo procedure
+#  aka 'ETCCDI' indix
 #
-# - temperature above 5°C on 5 consecutive days
+# first span of at least 6 consecutive days with daily mean temperature TG > 5°C
+# and first span after July 1st of 6 consecutive days with TG < 5°C.
 #
 # Reference:
-#  e.g.  FORMAYER, H., HAAS, P., HOFSTÄTTER, M., RADANOVICS, S. & KROMP-KOLB,
-#        H. (2007): Räumlich und zeitlich hochaufgelöste Temperaturszenarien
-#        für Wien und ausgewählte Analysen bezüglich Adaptionsstrategien.
-#        – BOKU-Met Bericht, 82 S.
+#  often known as Growing season Length (GSL) e.g.
+#   - Definition recommended by the CCl/CLIVAR/JCOMM Expert Team on Climate
+#     Change Detection and Indices (ETCCDI)
+#     http://etccdi.pacificclimate.org/list_27_indices.shtml
+#     http://www.climdex.org/indices.html
+#   - European Climate Assessment (ECA)
+#     http://eca.knmi.nl/indicesextremes/indicesdictionary.php
+#   - Frich, P. et al. (2002): Observed coherent changes in climatic extremes
+#     during the second half of the twentieth century. Climate Research (19):
+#     193-212.   http://www.climateknowledge.org/heat_waves/
+#     Doc2004_Frich_Extremes_Index_ClimResearch_2002.pdf
 #==============================================================================
-.start_std_climatology <- function(df, Tmin=5){
+.start_std_meteo <- function(df, Tmin=5){
+  # Assumptions:
+  # - data.frame 'df' contains month, DOY, Tavg
+  # - full years
 
   # mark days warmer than 5°C
   df$period <- ifelse(df$Tavg > Tmin, 1, 0)
 
-  # find first five day streak per year
-  years <- unique(df$year)
-  start <- integer(length(years))
-  for(i in 1:length(years)){
-    temp <- df[df$year == years[i], ]
-    temp$five <-  as.numeric(filter(temp$period, rep(1, 5), sides=1))
-    possible.start <- temp[!is.na(temp$five) & temp$five == 5, "DOY"]
-    if(length(possible.start) > 0)
-      start[i] <- min(possible.start)
-  }
+  # find first six day span per year
+  start <- tapply(df$period, df$year, FUN=function(x){
+    sixer <- as.numeric(filter(x, rep(1, 6), sides=1))
+    doy <- which(!is.na(sixer) & sixer == 6)
+    ifelse(length(doy) == 0, NA, min(doy))
+  })
 
   return(start)
 }