+++ R/vegperiod-package.R
@@ -3,8 +3,8 @@
 #' \tabular{ll}{
 #' Package: \tab vegperiod\cr
 #' Type: \tab Package\cr
-#' Version: \tab 0.1.0\cr
-#' Date: \tab 2015-04-06\cr
+#' Version: \tab 0.2.0\cr
+#' Date: \tab 2015-05-30\cr
 #' Depends: \tab R (>= 3.1.0)\cr
 #' Encoding: \tab UTF-8\cr
 #' License: \tab GPL (>= 3)\cr
@@ -12,12 +12,12 @@
 #' URL: \tab http://computerfoerster.de/r-pkgs\cr
 #' }
 #'
-#' Calculate start and end date of vegetation periods based on daily average air temperature and the day of the year (DOY).
+#' Calculate start and end date of vegetation period based on daily average air temperature and the day of the year (DOY). Functions for summing day degrees within vegetation period and reading freely available meteo data from DWD's ftp-Server are provided on top.
 #'
 #' @aliases vegperiod-package
 #' @name vegperiod-package
 #' @docType package
-#' @title The vegperiod Package
+#' @title Determine Vegetation Period of Forest Trees
 #' @author Robert Nuske
 #' @keywords package
 NULL